Remove a remaining old-python-pipeline task from contrib/python.
Move the new version of that task from tasks2 to tasks, and leave
a deprecation in the old module.

Also rename a target that had a non-canonical name for no good reason.
