+++ contrib/python/src/python/pants/contrib/python/checks/BUILD
@@ -5,7 +5,7 @@ contrib_plugin(
   name='plugin',
   dependencies=[
     'contrib/python/src/python/pants/contrib/python/checks/tasks/checkstyle:all',
-    'contrib/python/src/python/pants/contrib/python/checks/tasks:python',
+    'contrib/python/src/python/pants/contrib/python/checks/tasks',
     'contrib/python/src/python/pants/contrib/python/checks/tasks2',
     'src/python/pants/goal:task_registrar',
   ],
