+++ contrib/python/src/python/pants/contrib/python/checks/register.py
@@ -7,7 +7,7 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 
 from pants.goal.task_registrar import TaskRegistrar as task
 
-from pants.contrib.python.checks.tasks2.python_eval import PythonEval
+from pants.contrib.python.checks.tasks.python_eval import PythonEval
 from pants.contrib.python.checks.tasks.checkstyle.checker import PythonCheckStyleTask
