+++ contrib/python/src/python/pants/contrib/python/checks/tasks/BUILD
@@ -2,7 +2,6 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 python_library(
-  name='python',
   dependencies=[
     ':resources',
     'src/python/pants/backend/python/targets:python',
