+++ contrib/python/tests/python/pants_test/contrib/python/checks/tasks/BUILD
@@ -5,7 +5,9 @@ python_tests(
   name='python_eval',
   sources=['test_python_eval.py'],
   dependencies=[
-    'contrib/python/src/python/pants/contrib/python/checks/tasks:python',
+    'src/python/pants/backend/python/subsystems',
+    'contrib/python/src/python/pants/contrib/python/checks/tasks',
+    'src/python/pants/python',
     'tests/python/pants_test/backend/python/tasks:python_task_test_base',
   ]
 )
